templates/requirements.txt
# Cloud Run 배치 서비스 최소 의존성
# flask, gunicorn은 Dockerfile에서 별도 설치 (버전 고정)

# Google Cloud
google-cloud-bigquery>=3.0.0
google-cloud-secret-manager>=2.0.0

# HTTP (sync 잡에서 사용)
requests>=2.28.0

# 필요 시 추가:
# google-cloud-bigquery-storage>=2.0.0  # 대량 데이터 다운로드
# httpx>=0.24.0                         # async HTTP
# pandas>=2.0.0                         # 데이터 처리
# numpy>=2.0.0                          # 수치 연산
